+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * feat: Recipes and prebuilt canisters are now cached locally
 * feat: `icp canister migrate-id` - initiate canister ID migration across subnets
+* feat: install proxy canister when starting managed networks with all identities as controllers (or anonymous + default if more than 10 identities)
 
 # v0.1.0